moduls for main.js by specification
