+++ js/storage.js
@@ -1,0 +1,30 @@
+const GALLERY_LENGTH = 25;
+const DESCRIPTIONS = [
+  'спортзал',
+  'видеокарта',
+  'компьютер',
+  'солнышко',
+  'древесный кенгуру',
+  'хомячки',
+  'мыши',
+  'поляна',
+  'деревце',
+  'калина',
+  'верблюд',
+  'котик',
+  'машины',
+  'горы',
+  'мотоциклы',
+  'стол',
+  'розетка',
+  'фронтенд разработка',
+  'бэкенд разработчик',
+  'язык С или С++',
+  'микроконтроллер',
+  'STM32',
+  'Induction motor',
+  'ACED2021',
+  'IEEE',
+];
+
+export {GALLERY_LENGTH,DESCRIPTIONS};
